requirements.txt
oidcmsg>=1.3.0
pyyaml
jinja2>=2.11.3
responses>=0.13.0
